fix(vaults): poll requests only when events disconnect
